+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9V948B0DKXG2HEYKJ5QVP6H"
+        "01K9W3QETSRSTY22PT9Q578VWT"
     )
     r"""The unique identifier of the data source"""