+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -2130,6 +2130,7 @@ class CreatePromptPromptConfigTypedDict(TypedDict):
     messages: List[CreatePromptPromptsResponseMessagesTypedDict]
     stream: NotRequired[bool]
     model: NotRequired[str]
+    display_name: NotRequired[str]
     model_db_id: NotRequired[Nullable[str]]
     r"""The id of the resource"""
     model_type: NotRequired[Nullable[CreatePromptModelType]]
@@ -2150,6 +2151,8 @@ class CreatePromptPromptConfig(BaseModel):
     stream: Optional[bool] = None
 
     model: Optional[str] = None
+
+    display_name: Optional[str] = None
 
     model_db_id: OptionalNullable[str] = UNSET
     r"""The id of the resource"""
@@ -2172,6 +2175,7 @@ class CreatePromptPromptConfig(BaseModel):
         optional_fields = [
             "stream",
             "model",
+            "display_name",
             "model_db_id",
             "model_type",
             "model_parameters",