+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -636,6 +636,7 @@ class GetOnePromptPromptConfigTypedDict(TypedDict):
     messages: List[GetOnePromptMessagesTypedDict]
     stream: NotRequired[bool]
     model: NotRequired[str]
+    display_name: NotRequired[str]
     model_db_id: NotRequired[Nullable[str]]
     r"""The id of the resource"""
     model_type: NotRequired[Nullable[GetOnePromptModelType]]
@@ -656,6 +657,8 @@ class GetOnePromptPromptConfig(BaseModel):
     stream: Optional[bool] = None
 
     model: Optional[str] = None
+
+    display_name: Optional[str] = None
 
     model_db_id: OptionalNullable[str] = UNSET
     r"""The id of the resource"""
@@ -678,6 +681,7 @@ class GetOnePromptPromptConfig(BaseModel):
         optional_fields = [
             "stream",
             "model",
+            "display_name",
             "model_db_id",
             "model_type",
             "model_parameters",