+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9V948AXYR2QF4ZWQSDAHF3T"
+        "01K9W3QETM43AKQESSC3HA4Z53"
     )
     r"""The unique identifier of the data source"""