+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -666,6 +666,7 @@ class ListPromptVersionsPromptConfigTypedDict(TypedDict):
     messages: List[ListPromptVersionsMessagesTypedDict]
     stream: NotRequired[bool]
     model: NotRequired[str]
+    display_name: NotRequired[str]
     model_db_id: NotRequired[Nullable[str]]
     r"""The id of the resource"""
     model_type: NotRequired[Nullable[ListPromptVersionsModelType]]
@@ -686,6 +687,8 @@ class ListPromptVersionsPromptConfig(BaseModel):
     stream: Optional[bool] = None
 
     model: Optional[str] = None
+
+    display_name: Optional[str] = None
 
     model_db_id: OptionalNullable[str] = UNSET
     r"""The id of the resource"""
@@ -708,6 +711,7 @@ class ListPromptVersionsPromptConfig(BaseModel):
         optional_fields = [
             "stream",
             "model",
+            "display_name",
             "model_db_id",
             "model_type",
             "model_parameters",