+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9V948AZAYNR3WQG2QE1A6RS"
+        "01K9W3QETQMP1JBHMGNYAE4Q63"
     )
     r"""The unique identifier of the data source"""