+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K9V948B4E3GDBGD6E47TFV7M"
+        "01K9W3QETXNCF444490QQPB62D"
     )
     r"""The unique identifier of the data source"""