+++ packages/orq-rc/RELEASES.md
@@ -3949,3 +3949,13 @@ Based on:
 - [python v4.0.0-rc.26] packages/orq-rc
 ### Releases
 - [PyPI v4.0.0-rc.26] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.26 - packages/orq-rc
+
+## 2025-11-12 13:24:27
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.654.0 (2.750.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.0.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v4.0.0-rc.25] https://pypi.org/project/orq-ai-sdk/4.0.0-rc.25 - packages/orq-rc